refactor: extension, protocol 위치 변경
extension파일로~
