+++ BoxOffice/Model/Extension.swift
@@ -4,6 +4,15 @@
 //
 //  Created by Hisop on 2023/11/28.
 //
+import Foundation
+
+protocol URLSessionProtocol {
+    func dataTask(with url: URL,
+                  completionHandler: @escaping @Sendable (Data?, URLResponse?, Error?) -> Void)
+    -> URLSessionDataTask
+}
+
+extension URLSession: URLSessionProtocol { }
 
 extension BoxOffice: Equatable {
     static func == (lhs: BoxOffice, rhs: BoxOffice) -> Bool {
